Step 1
Add a space for the person doing the exercise to enter their First Name, Last Name, and Email Address. Please make this something they have to complete to do the exercise.

Step 2

Can we have an error show if they select more than 25, similar to the error in Step 4

Step 3

The maximum number to select should be 15 values
Can we have an error show if they select more than 15, similar to the error in Step 4

Step 5

If possible, can the order in which they select the values in Step 4 be shown first? If this is not possible or too much to do, don’t worry about it.
If this can’t be done, let's remove the sentence in Step 4 that says “Select 5 values in he order of importance.”